+++ Source/GTT/Public/Core/GTTGameplayStatics.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Kismet/BlueprintFunctionLibrary.h"
+#include "GTTGameplayStatics.generated.h"
+
+class APawn;
+class UGTTWantedComponent;
+
+UCLASS()
+class GTT_API UGTTGameplayStatics : public UBlueprintFunctionLibrary
+{
+    GENERATED_BODY()
+
+public:
+    UFUNCTION(BlueprintPure, Category="GTT|Wanted")
+    static UGTTWantedComponent* FindWantedComponentForPawn(APawn* Pawn);
+
+    UFUNCTION(BlueprintPure, Category="GTT|Wanted", meta=(WorldContext="WorldContextObject"))
+    static int32 GetPlayerWantedLevel(const UObject* WorldContextObject, int32 PlayerIndex = 0);
+};