+++ Source/GTT/Public/Police/GTTPolicePawn.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/Character.h"
+#include "GTTPolicePawn.generated.h"
+
+UCLASS()
+class GTT_API AGTTPolicePawn : public ACharacter
+{
+    GENERATED_BODY()
+
+public:
+    AGTTPolicePawn();
+};